New / "old" speller test result
This is now back to the old speller (ie the baseline one, close to the released speller in use), but with the exact same, cleaned test data as used with the new error model.

We should use this as comparison for the new ones, both regex and ML error models.